+++ lib/HomePage/filter.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_rating_bar/flutter_rating_bar.dart';
-import 'main.dart';
+import '../main.dart';
 
 void main() {
   runApp(new MyApp());
