+++ lib/OldFlavorMain/avg_rating.dart
